Code cleanup and significant performance improvements
